[SIX-179] refactor : develop에 새로 생성된 모듈로 변경
