+++ onedayhero-application/build.gradle
@@ -26,7 +26,6 @@ dependencies {
 
     // PROJECT
     implementation project(':onedayhero-common')
-    implementation project(':onedayhero-mongodb')
     implementation project(':onedayhero-domain')
     implementation project(':onedayhero-chat')
     implementation project(':onedayhero-mongo')
