+++ onedayhero-mongo/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
+    api 'org.springframework.boot:spring-boot-starter-data-mongodb'
 
     implementation project(':onedayhero-common')
 }
