+++ onedayhero-mongo/src/main/java/com/sixheroes/onedayheromongo/alarm/mongo/AlarmRepository.java
@@ -1,6 +1,6 @@
-package com.sixheroes.onedayheromongodb.alarm.mongo;
+package com.sixheroes.onedayheromongo.alarm.mongo;
 
-import com.sixheroes.onedayheromongodb.alarm.Alarm;
+import com.sixheroes.onedayheromongo.alarm.Alarm;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface AlarmRepository extends MongoRepository<Alarm, String> {
