--- onedayhero-mongo/src/main/java/com/sixheroes/onedayheromongo/service/alarm/AlarmService.java
@@ -1,4 +1,0 @@
-package com.sixheroes.onedayheromongo.service.alarm;
-
-public class AlarmService {
-}
