+++ onedayhero-mongo/src/test/java/com/sixheroes/onedayheromongo/alarm/mongo/AlarmTemplateRepositoryTest.java
@@ -1,14 +1,16 @@
-package com.sixheroes.onedayheromongodb.alarm.mongo;
+package com.sixheroes.onedayheromongo.alarm.mongo;
 
-import com.sixheroes.onedayheromongodb.IntegrationMongoRepositoryTest;
-import com.sixheroes.onedayheromongodb.alarm.AlarmTemplate;
+import com.sixheroes.onedayheromongo.IntegrationMongoRepositoryTest;
+import com.sixheroes.onedayheromongo.alarm.AlarmTemplate;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@Disabled
 class AlarmTemplateRepositoryTest extends IntegrationMongoRepositoryTest {
 
     @Autowired
