--- onedayhero-mongodb/build.gradle
@@ -1,17 +1,0 @@
-dependencies {
-	api 'org.springframework.boot:spring-boot-starter-data-mongodb'
-
-	implementation project(':onedayhero-common');
-}
-
-tasks.named('test') {
-	useJUnitPlatform()
-}
-
-jar {
-	enabled = true
-}
-
-bootJar {
-	enabled = false
-}
