--- onedayhero-mongodb/src/test/java/com/sixheroes/OnedayheroDomainMongodbApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sixheroes;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-class OnedayheroDomainMongodbApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
